+++ packages/flutter_universal_sync_core/test/contract_suite_test.dart
@@ -1,8 +1,6 @@
 import 'package:flutter_universal_sync_core/flutter_universal_sync_core.dart';
 import 'package:flutter_universal_sync_core/testing.dart';
 
-import 'support/in_memory_adapter.dart';
-
 void main() {
   runLocalDatabaseAdapterContract(
     factory: InMemoryAdapter.new,